+++ src/main/java/com/ward/ward_server/api/item/controller/PublicItemController.java
@@ -1,7 +1,6 @@
 package com.ward.ward_server.api.item.controller;
 
 import com.ward.ward_server.api.item.dto.ItemDetailResponse;
-import com.ward.ward_server.api.item.dto.ItemRequest;
 import com.ward.ward_server.api.item.dto.ItemSimpleResponse;
 import com.ward.ward_server.api.item.dto.ItemTopRankResponse;
 import com.ward.ward_server.api.item.entity.enums.Category;
@@ -24,24 +23,11 @@ import static com.ward.ward_server.global.response.ApiResponseMessage.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/items")
-public class ItemController {
+@RequestMapping("/public/items")
+public class PublicItemController {
 
     private final ItemService itemService;
     private final ItemViewCountScheduler itemViewCountScheduler;
-
-    @PostMapping("/execute-update-view-counts")
-    public ApiResponse<String> abc() {
-        itemViewCountScheduler.executeUpdateViewCounts();
-        return ApiResponse.ok("성공");
-    }
-
-
-    @PostMapping
-    public ApiResponse<ItemDetailResponse> createItem(@RequestBody ItemRequest request) throws ApiException {
-        return ApiResponse.ok(ITEM_CREATE_SUCCESS,
-                itemService.createItem(request.itemCode(), request.koreanName(), request.englishName(), request.mainImage(), request.itemImages(), request.brandId(), request.category(), request.price()));
-    }
 
     @GetMapping("/{itemId}/details")
     public ApiResponse<ItemDetailResponse> getItem(@PathVariable("itemId") Long itemId) {
@@ -73,18 +59,5 @@ public class ItemController {
 
         List<ItemTopRankResponse> topItemsResponse = itemService.getTopItemsResponseByCategory(category, limit);
         return ApiResponse.ok(REALTIME_TOP_LOAD_SUCCESS, topItemsResponse);
-    }
-
-    @PatchMapping("/{itemId}")
-    public ApiResponse<ItemDetailResponse> updateItem(@PathVariable("itemId") Long itemId,
-                                                      @RequestBody ItemRequest request) {
-        return ApiResponse.ok(ITEM_UPDATE_SUCCESS,
-                itemService.updateItem(itemId, request.koreanName(), request.englishName(), request.itemCode(), request.mainImage(), request.itemImages(), request.brandId(), request.category(), request.price()));
-    }
-
-    @DeleteMapping("/{itemId}")
-    public ApiResponse<Void> deleteItem(@PathVariable("itemId") Long itemId) {
-        itemService.deleteItem(itemId);
-        return ApiResponse.ok(ITEM_DELETE_SUCCESS);
     }
 }
