+++ src/main/java/com/ward/ward_server/api/releaseInfo/controller/AdminReleaseInfoController.java
@@ -1,0 +1,59 @@
+package com.ward.ward_server.api.releaseInfo.controller;
+
+import com.ward.ward_server.api.releaseInfo.dto.ReleaseInfoDetailResponse;
+import com.ward.ward_server.api.releaseInfo.dto.ReleaseInfoRequest;
+import com.ward.ward_server.api.releaseInfo.service.ReleaseInfoService;
+import com.ward.ward_server.global.response.ApiResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.*;
+
+import static com.ward.ward_server.global.response.ApiResponseMessage.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/admin/release-infos")
+@Slf4j
+public class AdminReleaseInfoController {
+    private final ReleaseInfoService releaseInfoService;
+
+    @PostMapping
+    public ApiResponse<ReleaseInfoDetailResponse> createReleaseInfo(@RequestBody ReleaseInfoRequest request) {
+        return ApiResponse.ok(RELEASE_INFO_CREATE_SUCCESS, releaseInfoService.createReleaseInfo(
+                request.itemId(),
+                request.platformName(),
+                request.siteUrl(),
+                request.releaseDate(),
+                request.dueDate(),
+                request.presentationDate(),
+                request.releasePrice(),
+                request.currencyUnit(),
+                request.notificationMethod(),
+                request.releaseMethod(),
+                request.deliveryMethod()));
+    }
+
+    @PatchMapping("/{releaseInfoId}")
+    public ApiResponse<ReleaseInfoDetailResponse> updateReleaseInfo(@PathVariable("releaseInfoId") Long releaseInfoId,
+                                                                    @RequestBody ReleaseInfoRequest request) {
+        return ApiResponse.ok(RELEASE_INFO_UPDATE_SUCCESS, releaseInfoService.updateReleaseInfo(
+                releaseInfoId,
+                request.itemId(),
+                request.platformName(),
+                request.siteUrl(),
+                request.releaseDate(),
+                request.dueDate(),
+                request.presentationDate(),
+                request.releasePrice(),
+                request.currencyUnit(),
+                request.notificationMethod(),
+                request.releaseMethod(),
+                request.deliveryMethod()));
+    }
+
+    @DeleteMapping("/{releaseInfoId}")
+    public ApiResponse<Void> deleteReleaseInfo(@PathVariable("releaseInfoId") Long releaseInfoId) {
+        releaseInfoService.deleteReleaseInfo(releaseInfoId);
+        return ApiResponse.ok(RELEASE_INFO_DELETE_SUCCESS);
+    }
+}
