+++ src/main/java/com/ward/ward_server/api/search/controller/SearchController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/search")
+@RequestMapping("/public/search")
 public class SearchController {
 
     private final SearchService searchService;
